refactor: removed logic for drawing in feature-map
